Move two of the logic-heavy lane transformations into raw_map.
This temporarily breaks a few tests in other places, porting that code
over momentarily...
