+++ map_model/src/edits/mod.rs
@@ -14,10 +14,9 @@ use raw_map::{get_lane_specs_ltr, InputRoad};
 pub use self::perma::PermanentMapEdits;
 use crate::make::{match_points_to_lanes, snap_driveway, trim_path};
 use crate::{
-    connectivity, AccessRestrictions, BuildingID, ControlStopSign, ControlTrafficSignal, Direction,
-    IntersectionID, IntersectionType, LaneID, LaneSpec, LaneType, Map, MapConfig, Movement,
-    ParkingLotID, PathConstraints, Pathfinder, Road, RoadID, TransitRouteID, TurnID, TurnType,
-    Zone,
+    connectivity, AccessRestrictions, BuildingID, ControlStopSign, ControlTrafficSignal,
+    IntersectionID, IntersectionType, LaneID, LaneSpec, Map, MapConfig, Movement, ParkingLotID,
+    PathConstraints, Pathfinder, Road, RoadID, TransitRouteID, TurnID, TurnType, Zone,
 };
 
 mod compat;
@@ -118,64 +117,6 @@ impl EditRoad {
             changes.push("access restrictions".to_string());
         }
         changes
-    }
-
-    /// Transforms a string describing lane types and directions, like "spddps" and "vv^^^^^", into
-    /// an EditRoad. Useful for unit tests.
-    pub fn create_for_test(input_lt: &str, input_dir: &str) -> EditRoad {
-        assert_eq!(input_lt.len(), input_dir.len());
-        EditRoad {
-            lanes_ltr: input_lt
-                .chars()
-                .zip(input_dir.chars())
-                .map(|(lt, dir)| LaneSpec {
-                    lt: LaneType::from_char(lt),
-                    dir: if dir == '^' {
-                        Direction::Fwd
-                    } else {
-                        Direction::Back
-                    },
-                    // Dummy
-                    width: Distance::ZERO,
-                })
-                .collect(),
-            speed_limit: Speed::ZERO,
-            access_restrictions: AccessRestrictions::new(),
-        }
-    }
-
-    /// This is meant for table-driven unit tests. Call this on the transformed / output lanes. If
-    /// the lanes don't match, `ok` will be set to false and appropriate errors will be printed.
-    pub fn check_lanes_ltr(
-        &self,
-        description: String,
-        input_lt: &str,
-        input_dir: &str,
-        expected_lt: &str,
-        expected_dir: &str,
-        ok: &mut bool,
-    ) {
-        let actual_lt: String = self.lanes_ltr.iter().map(|s| s.lt.to_char()).collect();
-        let actual_dir: String = self
-            .lanes_ltr
-            .iter()
-            .map(|s| if s.dir == Direction::Fwd { '^' } else { 'v' })
-            .collect();
-
-        if actual_lt != expected_lt || actual_dir != expected_dir {
-            *ok = false;
-            println!("{}", description);
-            println!("Input:");
-            println!("    {}", input_lt);
-            println!("    {}", input_dir);
-            println!("Got:");
-            println!("    {}", actual_lt);
-            println!("    {}", actual_dir);
-            println!("Expected:");
-            println!("    {}", expected_lt);
-            println!("    {}", expected_dir);
-            println!();
-        }
     }
 }
