+++ raw_map/src/lib.rs
@@ -25,6 +25,7 @@ pub use self::types::{
     LaneType, MapConfig, NamePerLanguage, NORMAL_LANE_THICKNESS, SIDEWALK_THICKNESS,
 };
 
+mod edit;
 mod export;
 pub mod geometry;
 pub mod initial;
